requirements.txt
# Web Framework
Django==5.0.6  # Using the latest version you have listed
django-widget-tweaks  # Added as per your mention, version not specified

# Database
dj_database_url==2.2.0

# Image Processing
Pillow==11.0.0
ImageHash==4.3.1
opencv_python_headless==4.10.0.84
rembg

# Data Processing and Analysis
numpy==2.1.2
pandas==2.2.3
scipy==1.14.1
pytz==2024.1
python_dateutil==2.9.0.post0

# Machine Learning
scikit_learn==1.5.2
tensorflow==2.17.0

# File Handling
openpyxl==3.1.5
XlsxWriter==3.2.0

# PDF Generation
reportlab==4.2.2

# Authentication
social_auth_app_django==5.4.1
social_auth_core==4.5.4

# Payment Processing
stripe==10.7.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.5
webcolors==24.8.0
